Added level loading system and movement to square

+++ src/math/vector4.cpp
@@ -11,11 +11,11 @@ Vector4 Vector4::operator-(const Vector4& Y) const{
   return Vector4(this->X - Y.X, this->Y - Y.Y, this->Z - Y.Z, this->W - Y.W);
 }
 
-Vector4 Vector4::operator*(const double& Scalor) const{
+Vector4 Vector4::operator*(const float& Scalor) const{
   return Vector4(this->X * Scalor, this->Y * Scalor, this->Z * Scalor, this->W * Scalor);
 }
 
-Vector4 Vector4::operator/(const double& Scalor) const{
+Vector4 Vector4::operator/(const float& Scalor) const{
   return Vector4(this->X / Scalor, this->Y / Scalor, this->Z / Scalor, this->W / Scalor);
 }
