+++ src/math/vector4.h
@@ -17,17 +17,17 @@ public:
 public:
   Vector4()
     :X(0.0f), Y(0.0f), Z(0.0f), W(0.0f) {}
-  Vector4(double x, double y, double z, double w)
+  Vector4(float x, float y, float z, float w)
     :X(x), Y(y), Z(z), W(w) {}
-  Vector4(double x)
+  Vector4(float x)
     :X(x), Y(x), Z(x), W(x) {}
 
   void Print() const;
 
   Vector4 operator+(const Vector4& Y) const;
   Vector4 operator-(const Vector4& Y) const;
-  Vector4 operator*(const double& Scalor) const;
-  Vector4 operator/(const double& Scalor) const;
+  Vector4 operator*(const float& Scalor) const;
+  Vector4 operator/(const float& Scalor) const;
 };
 
 #endif
